requirements.txt
whisper
ffmpeg
torch
yfinance
spacy
textblob
haystack
streamlit
pdfplumber 
requests 
beautifulsoup4 
scikit-learn 
statsmodels 
transformers
duckduckgo-search 
google-generativeai
